requirements.txt
astroid==2.2.5
et-xmlfile==1.0.1
inflect==2.1.0
isort==4.3.21
jdcal==1.4.1
lazy-object-proxy==1.4.2
mccabe==0.6.1
openpyxl==2.6.3
pkg-resources==0.0.0
psycopg2-binary==2.8.3
pylint==2.3.1
rope==0.14.0
six==1.12.0
sqlacodegen==2.1.0
SQLAlchemy==1.3.8
typed-ast==1.4.0
wrapt==1.11.2
